+++ ktlint-ruleset-standard/src/main/kotlin/com/github/shyiko/ktlint/ruleset/standard/NoSemicolonsRule.kt
@@ -2,10 +2,12 @@ package com.github.shyiko.ktlint.ruleset.standard
 
 import com.github.shyiko.ktlint.core.Rule
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
+import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.PsiWhiteSpaceImpl
 import org.jetbrains.kotlin.com.intellij.psi.util.PsiTreeUtil
+import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.KtEnumEntry
 
 class NoSemicolonsRule : Rule("no-semi") {
@@ -18,15 +20,20 @@ class NoSemicolonsRule : Rule("no-semi") {
         if (node is LeafPsiElement && node.textMatches(";") && !node.isPartOfString() &&
                 !node.isPartOf(KtEnumEntry::class)) {
             val nextLeaf = PsiTreeUtil.nextLeaf(node, true)
-            if (nextLeaf == null /* eof */ ||
-                (nextLeaf is PsiWhiteSpace && (nextLeaf.text.contains("\n") ||
-                    PsiTreeUtil.nextLeaf(nextLeaf, true) == null /* \s+ and then eof */))
-                ) {
+            if (doesNotRequirePreSemi(nextLeaf)) {
+                if (node.psi.prevLeafIgnoringWhitespaceAndComments()?.node?.elementType == KtTokens.OBJECT_KEYWORD) {
+                    // https://github.com/shyiko/ktlint/issues/281
+                    return
+                }
                 emit(node.startOffset, "Unnecessary semicolon", true)
                 if (autoCorrect) {
                     node.treeParent.removeChild(node)
                 }
             } else if (nextLeaf !is PsiWhiteSpace) {
+                val prevLeaf = PsiTreeUtil.prevLeaf(node, true)
+                if (prevLeaf is PsiWhiteSpace && prevLeaf.textContains('\n')) { // \n;{
+                    return
+                }
                 // todo: move to a separate rule
                 emit(node.startOffset + 1, "Missing spacing after \";\"", true)
                 if (autoCorrect) {
@@ -34,5 +41,16 @@ class NoSemicolonsRule : Rule("no-semi") {
                 }
             }
         }
+    }
+
+    private fun doesNotRequirePreSemi(nextLeaf: PsiElement?): Boolean {
+        if (nextLeaf is PsiWhiteSpace) {
+            val nextNextLeaf = PsiTreeUtil.nextLeaf(nextLeaf, true)
+            return (
+                nextNextLeaf == null || // \s+ and then eof
+                nextLeaf.textContains('\n') && !nextNextLeaf.textMatches("{")
+            )
+        }
+        return nextLeaf == null /* eof */
     }
 }