+++ ktlint-ruleset-standard/src/main/kotlin/com/github/shyiko/ktlint/ruleset/standard/ParameterListWrappingRule.kt
@@ -5,7 +5,9 @@ import org.jetbrains.kotlin.KtNodeTypes
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.lang.FileASTNode
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
+import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafElement
+import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.PsiWhiteSpaceImpl
 import org.jetbrains.kotlin.com.intellij.psi.util.PsiTreeUtil
 import org.jetbrains.kotlin.lexer.KtTokens
@@ -15,6 +17,7 @@ import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes
 class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
 
     private var indentSize = -1
+    private var maxLineLength = -1
 
     override fun visit(
         node: ASTNode,
@@ -24,6 +27,7 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
         if (node.elementType == KtStubElementTypes.FILE) {
             val ec = EditorConfig.from(node as FileASTNode)
             indentSize = ec.indentSize
+            maxLineLength = ec.maxLineLength
             return
         }
         if (indentSize <= 0) {
@@ -36,24 +40,37 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
             // - at least one of the parameters is
             // - maxLineLength exceeded (and separating parameters with \n would actually help)
             // in addition, "(" and ")" must be on separates line if any of the parameters are (otherwise on the same)
-            val putParametersOnSeparateLines = node.textContains('\n')
-                // maxLineLength > 0 && node.lineLength() > maxLineLength
+            val putParametersOnSeparateLines = node.textContains('\n') ||
+                // max_line_length exceeded
+                maxLineLength > -1 && (node.psi.column - 1 + node.textLength) > maxLineLength
             if (putParametersOnSeparateLines) {
                 // aiming for
                 // ... LPAR
-                // <LPAR line indent + indentSize> VALUE_PARAMETER...
-                // <LPAR line indent> RPAR
+                // <line indent + indentSize> VALUE_PARAMETER...
+                // <line indent> RPAR
                 val indent = "\n" + node.psi.lineIndent()
                 val paramIndent = indent + " ".repeat(indentSize) // single indent as recommended by Jetbrains/Google
                 nextChild@ for (child in node.children()) {
                     when (child.elementType) {
+                        KtTokens.LPAR -> {
+                            val prevLeaf = child.psi.prevLeaf()
+                            if (prevLeaf is PsiWhiteSpace && prevLeaf.textContains('\n')) {
+                                emit(child.startOffset, errorMessage(child), true)
+                                if (autoCorrect) {
+                                    prevLeaf.delete()
+                                }
+                            }
+                        }
                         KtStubElementTypes.VALUE_PARAMETER,
                         KtTokens.RPAR -> {
                             var paramInnerIndentAdjustment = 0
-                            val prevLeaf = child.psi.prevLeaf()!!
-                            val intendedIndent = if (child.elementType == KtStubElementTypes.VALUE_PARAMETER)
-                                paramIndent else indent
-                            if (prevLeaf.node.elementType == KtTokens.WHITE_SPACE) {
+                            val prevLeaf = child.psi.prevLeaf()
+                            val intendedIndent = if (child.elementType == KtStubElementTypes.VALUE_PARAMETER) {
+                                paramIndent
+                            } else {
+                                indent
+                            }
+                            if (prevLeaf is PsiWhiteSpace) {
                                 val spacing = prevLeaf.text
                                 val cut = spacing.lastIndexOf("\n")
                                 if (cut > -1) {
@@ -69,7 +86,7 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
                                 if (autoCorrect) {
                                     val adjustedIndent = (if (cut > -1) spacing.substring(0, cut) else "") + intendedIndent
                                     paramInnerIndentAdjustment = adjustedIndent.length - prevLeaf.textLength
-                                    prevLeaf.rawReplaceWithText(adjustedIndent)
+                                    (prevLeaf as LeafPsiElement).rawReplaceWithText(adjustedIndent)
                                 }
                             } else {
                                 emit(child.startOffset, errorMessage(child), true)
@@ -115,13 +132,9 @@ class ParameterListWrappingRule : Rule("parameter-list-wrapping") {
             return offsetToTheLeft + 1
         }
 
-    private fun ASTNode.visit(cb: (node: ASTNode) -> Unit) {
-        cb(this)
-        this.getChildren(null).forEach { it.visit(cb) }
-    }
-
     private fun errorMessage(node: ASTNode) =
         when (node.elementType) {
+            KtTokens.LPAR -> """Unnecessary newline before "(""""
             KtStubElementTypes.VALUE_PARAMETER ->
                 "Parameter should be on a separate line (unless all parameters can fit a single line)"
             KtTokens.RPAR -> """Missing newline before ")""""