+++ ktlint-ruleset-standard/src/main/kotlin/com/github/shyiko/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -7,7 +7,13 @@ class StandardRuleSetProvider : RuleSetProvider {
 
     override fun get(): RuleSet = RuleSet("standard",
         ChainWrappingRule(),
+        CommentSpacingRule(),
+        FilenameRule(),
         FinalNewlineRule(),
+        PackageNameRule(),
+        // disabled until auto-correct is working properly
+        // (e.g. try formatting "if (true)\n    return { _ ->\n        _\n}")
+        // MultiLineIfElseRule(),
         // disabled until it's clear how to reconcile difference in Intellij & Android Studio import layout
         // ImportOrderingRule(),
         IndentationRule(),
@@ -32,6 +38,7 @@ class StandardRuleSetProvider : RuleSetProvider {
         SpacingAroundCurlyRule(),
         SpacingAroundKeywordRule(),
         SpacingAroundOperatorsRule(),
+        SpacingAroundParensRule(),
         SpacingAroundRangeOperatorRule(),
         StringTemplateRule()
     )