+++ ktlint-ruleset-standard/src/test/kotlin/com/github/shyiko/ktlint/ruleset/standard/NoSemicolonsRuleTest.kt
@@ -18,6 +18,12 @@ class NoSemicolonsRuleTest {
                 fun name() { a(); return b }
                 println(";")
                 println();
+
+                Any();
+                {
+                }.print()
+                Any()
+                ;{ /*...*/ }.print()
             }
             """.trimIndent()
         )).isEqualTo(listOf(
@@ -62,5 +68,25 @@ class NoSemicolonsRuleTest {
             }
             """.trimIndent()
         )
+    }
+
+    @Test
+    fun testSemiIsPreservedAfterCompanionObject() {
+        // github issue #281
+        assertThat(NoSemicolonsRule().lint(
+            """
+            class A {
+                companion object;
+                companion object ;
+            }
+            class A {
+                companion object {
+                    val s = ""
+                };
+            }
+            """.trimIndent()
+        )).isEqualTo(listOf(
+            LintError(8, 6, "no-semi", "Unnecessary semicolon")
+        ))
     }
 }